+++ app/build.gradle
@@ -50,13 +50,16 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
 
     //Retrofit
-    implementation 'com.squareup.okhttp3:logging-interceptor:3.13.1'
-    implementation 'com.squareup.retrofit2:converter-gson:2.3.0'
-    implementation 'com.squareup.retrofit2:retrofit:2.6.1'
+    implementation 'com.squareup.okhttp3:logging-interceptor:4.8.0'
+    implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
+    implementation 'com.squareup.retrofit2:retrofit:2.9.0'
 
     //Data Binding
     kapt 'com.android.databinding:compiler:3.1.4'
 
     //Viewmodels
     implementation 'androidx.fragment:fragment-ktx:1.3.0'
+
+    //Network
+    implementation "com.github.LTMezzari:kotlin-network:0.8.3"
 }
